Create de_DE.xml
I made this translation fit to the translation of AE.
Any suggestions are appreciated!
I hope, this has to be be UTF-8 encoded, otherwise I will change it (If it is NOT UTF-8 but ISO-8859-1, then all äöüß will be shown strange ingame).

P.S.: The "K" at "256K Storage" shouldn't be capitalized. The "M" for "Mega" is capitalized correctly, though.