Block.Drain fix
After having a play with some custom fluids I noticed that none of the IFluidBlock's would work with the Annihilation Plane. It seems the wrong co-ordinates was being checked when using the 'block.drain' function. I have made and tested the changes and they seem to work. Please give them a try.