*Really* fix the OreDict exporter filter parsing now.
Pattern Syntax: [!][@~]<pattern>
Filter syntax: <pattern> [&|] <pattern2> [&|] ...
Parentheses NOT supported.
@ prefix for resource namespace matching (i.e. @minecraft @appliedenergistics)
~ prefix for resource path matching
! for negation (use before @ and ~ if applicable)
* can be used in any position, although usage of * that isn't beginning or end may result in performance hits due to usage of regex matching.

Usage with pure oredict matches (no @ and ~) will result in querying the oredict once and building a whitelist that is saved and the filter is only evaluated once.
Using @ and ~ will result in the filter being evaluated for every item in AE until it finds a match.

Examples tested:
@minecraft & ore* (All vanilla ores)
@mine* | @thermal* (All blocks from mods with namespaces that starts with mine or thermal)
oreGold | oreIron (Gold and iron ore only)