Fix fluid interface dupe
This dupe issue will happen when you attach two or more tanks to the same fluid interface and the fluid interface pushes some fluid as in some pattern

Note to future programmers: this pushItems() logic is not as clean as it can be and can push 1 stack per tick whereas vanilla can push all stacks at the same tick.
It deserves a complete rewrite.
I don't do this because I don't fully understand this piece of code.

Signed-off-by: Glease <4586901+Glease@users.noreply.github.com>